+++ audio-streaming-client-java/src/main/java/com/baidu/acu/pie/demo/JavaDemo.java
@@ -2,21 +2,20 @@
 
 package com.baidu.acu.pie.demo;
 
-import com.baidu.acu.pie.AudioStreaming.AudioFragmentRequest;
-import com.baidu.acu.pie.client.AsrClient;
-import com.baidu.acu.pie.client.AsrClientFactory;
-import com.baidu.acu.pie.client.Consumer;
-import com.baidu.acu.pie.model.AsrConfig;
-import com.baidu.acu.pie.model.AsrProduct;
-import com.baidu.acu.pie.model.RecognitionResult;
-import com.baidu.acu.pie.model.StreamContext;
-import com.google.protobuf.ByteString;
-import org.joda.time.DateTime;
-
 import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.List;
+
+import org.joda.time.DateTime;
+
+import com.baidu.acu.pie.client.AsrClient;
+import com.baidu.acu.pie.client.AsrClientFactory;
+import com.baidu.acu.pie.client.Consumer;
+import com.baidu.acu.pie.model.AsrConfig;
+import com.baidu.acu.pie.model.RecognitionResult;
+import com.baidu.acu.pie.model.RequestMetaData;
+import com.baidu.acu.pie.model.StreamContext;
 
 /**
  * JavaDemo
@@ -26,6 +25,7 @@ import java.util.List;
 public class JavaDemo {
     public static void main(String[] args) {
         JavaDemo javaDemo = new JavaDemo();
+        //        javaDemo.recognizeFile();
         javaDemo.asyncRecognition();
     }
 
@@ -35,8 +35,9 @@ public class JavaDemo {
                 .serverIp("127.0.0.1")
                 .serverPort(80)
                 .appName("simple demo")
-                .sleepRatio(1)
-                .product(AsrProduct.CUSTOMER_SERVICE_FINANCE);
+                .productId("1903")
+                .userName("user1")
+                .password("password1");
 
         return AsrClientFactory.buildClient(asrConfig);
     }
@@ -44,7 +45,10 @@ public class JavaDemo {
     public void recognizeFile() {
         String audioFilePath = "testaudio/bj8k.wav";
         AsrClient asrClient = createAsrClient();
-        List<RecognitionResult> results = asrClient.syncRecognize(Paths.get(audioFilePath));
+
+        RequestMetaData requestMetaData = RequestMetaData.defaultRequestMeta();
+        requestMetaData.enableFlushData(false);
+        List<RecognitionResult> results = asrClient.syncRecognize(Paths.get(audioFilePath).toFile(), requestMetaData);
 
         // don't forget to shutdown !!!
         asrClient.shutdown();
@@ -53,16 +57,46 @@ public class JavaDemo {
             System.out.println(String.format(AsrConfig.TITLE_FORMAT,
                     "file_name",
                     "serial_num",
-                    "err_no",
-                    "err_message",
                     "start_time",
                     "end_time",
                     "result"));
             System.out.println(String.format(AsrConfig.TITLE_FORMAT,
                     audioFilePath,
                     result.getSerialNum(),
-                    result.getErrorCode(),
-                    result.getErrorMessage(),
+                    result.getStartTime(),
+                    result.getEndTime(),
+                    result.getResult()
+            ));
+        }
+    }
+
+    /**
+     * 用户可以自己创建一个 RequestMeta 对象，用来控制请求时候的数据发送速度等参数
+     */
+    public void recognizeFileWithRequestMeta() {
+        String audioFilePath = "testaudio/bj8k.wav";
+        AsrClient asrClient = createAsrClient();
+
+        RequestMetaData requestMetaData = new RequestMetaData();
+        requestMetaData.sendPerSeconds(0.05);
+        requestMetaData.sendPackageRatio(1);
+        requestMetaData.sleepRatio(1);
+
+        List<RecognitionResult> results = asrClient.syncRecognize(Paths.get(audioFilePath).toFile(), requestMetaData);
+
+        // don't forget to shutdown !!!
+        asrClient.shutdown();
+
+        for (RecognitionResult result : results) {
+            System.out.println(String.format(AsrConfig.TITLE_FORMAT,
+                    "file_name",
+                    "serial_num",
+                    "start_time",
+                    "end_time",
+                    "result"));
+            System.out.println(String.format(AsrConfig.TITLE_FORMAT,
+                    audioFilePath,
+                    result.getSerialNum(),
                     result.getStartTime(),
                     result.getEndTime(),
                     result.getResult()
@@ -81,8 +115,8 @@ public class JavaDemo {
             @Override
             public void accept(RecognitionResult it) {
                 System.out.println(
-//                        Instant.now().toString() + "\t" + Thread.currentThread().getId() + " receive fragment: " + it);
-                        Thread.currentThread().getId() + " receive fragment: " + it);
+                        DateTime.now().toString() + "\t" + Thread.currentThread().getId() +
+                                " receive fragment: " + it);
             }
         });
 
@@ -95,17 +129,13 @@ public class JavaDemo {
 
             // 使用 sender.onNext 方法，将 InputStream 中的数据不断地发送到 asr 后端，发送的最小单位是 AudioFragment
             while ((readSize = audioStream.read(data)) != -1 && !streamContext.getFinishLatch().finished()) {
-                System.out.println(new DateTime().toString() + "\t"
-                        + Thread.currentThread().getId() + " send fragment");
-                streamContext.getSender().onNext(AudioFragmentRequest.newBuilder()
-                        .setAudioData(ByteString.copyFrom(data, 0, readSize))
-                        .build());
+                streamContext.send(data);
                 // 主动休眠一段时间，来模拟人说话场景下的音频产生速率
                 // 在对接麦克风等设备的时候，可以去掉这个 sleep
                 Thread.sleep(20);
             }
             System.out.println(new DateTime().toString() + "\t" + Thread.currentThread().getId() + " send finish");
-            streamContext.getSender().onCompleted();
+            streamContext.complete();
 
             // wait to ensure to receive the last response
             streamContext.getFinishLatch().await();