+++ ios-demo/BaiduARS/BaiduARS/AppDelegate.h
@@ -1,0 +1,17 @@
+//
+//  AppDelegate.h
+//  BaiduARS
+//
+//  Created by Wu,Yunpeng(AI2B) on 2019/4/12.
+//  Copyright © 2019 Cloud. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface AppDelegate : UIResponder <UIApplicationDelegate>
+
+@property (strong, nonatomic) UIWindow *window;
+
+
+@end
+