+++ ios-demo/BaiduARS/BaiduARS/main.m
@@ -1,0 +1,16 @@
+//
+//  main.m
+//  BaiduARS
+//
+//  Created by Wu,Yunpeng(AI2B) on 2019/4/12.
+//  Copyright © 2019 Cloud. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+#import "AppDelegate.h"
+
+int main(int argc, char * argv[]) {
+    @autoreleasepool {
+        return UIApplicationMain(argc, argv, nil, NSStringFromClass([AppDelegate class]));
+    }
+}